+++ toolsrc/include/vcpkg/export.chocolatey.h
@@ -1,7 +1,8 @@
 #pragma once
 
+#include <vcpkg/fwd/vcpkgpaths.h>
+
 #include <vcpkg/dependencies.h>
-#include <vcpkg/vcpkgpaths.h>
 
 #include <vector>
