+++ toolsrc/include/vcpkg/export.prefab.h
@@ -1,9 +1,10 @@
 #pragma once
 
+#include <vcpkg/fwd/vcpkgpaths.h>
+
 #include <vcpkg/base/system.h>
 
 #include <vcpkg/dependencies.h>
-#include <vcpkg/vcpkgpaths.h>
 
 #include <vector>
