+++ toolsrc/include/vcpkg/fwd/build.h
@@ -1,0 +1,7 @@
+#pragma once
+
+namespace vcpkg::Build
+{
+    struct BuildInfo;
+    struct PreBuildInfo;
+}
