+++ toolsrc/include/vcpkg/fwd/cmakevars.h
@@ -1,0 +1,6 @@
+#pragma once
+
+namespace vcpkg::CMakeVars
+{
+    struct CMakeVarProvider;
+}
