+++ toolsrc/include/vcpkg/fwd/dependencies.h
@@ -1,0 +1,7 @@
+#pragma once
+
+namespace vcpkg::Dependencies
+{
+    struct InstallPlanAction;
+    struct ActionPlan;
+}
