+++ toolsrc/include/vcpkg/fwd/portfileprovider.h
@@ -1,0 +1,7 @@
+#pragma once
+
+namespace vcpkg::PortFileProvider
+{
+    struct PortFileProvider;
+    struct PathsPortFileProvider;
+}
