+++ toolsrc/include/vcpkg/remove.h
@@ -1,14 +1,10 @@
 #pragma once
 
+#include <vcpkg/fwd/dependencies.h>
 #include <vcpkg/fwd/vcpkgcmdarguments.h>
 #include <vcpkg/fwd/vcpkgpaths.h>
 
 #include <vcpkg/commands.interface.h>
-
-namespace vcpkg::Dependencies
-{
-    struct RemovePlanAction;
-}
 
 namespace vcpkg::Remove
 {
