+++ toolsrc/src/vcpkg-test/commands.build.cpp
@@ -2,6 +2,7 @@
 
 #include <vcpkg/base/files.h>
 
+#include <vcpkg/build.h>
 #include <vcpkg/commands.h>
 #include <vcpkg/vcpkgcmdarguments.h>
 #include <vcpkg/vcpkgpaths.h>
