+++ toolsrc/src/vcpkg/input.cpp
@@ -5,6 +5,7 @@
 #include <vcpkg/input.h>
 #include <vcpkg/metrics.h>
 #include <vcpkg/packagespec.h>
+#include <vcpkg/vcpkgpaths.h>
 
 namespace vcpkg
 {
